修复一些角落编译警告；其他杂项改动

+++ src/lammp/numth/binvert_1.c
@@ -29,7 +29,7 @@ static const uchar binv_tab[128] = {
 
 uint lmmp_binvert_uint_(uint a) {
     lmmp_param_assert(a % 2 == 1);
-    ulong r, y;
+    uint r, y;
 
     r = binv_tab[(a / 2) & 0x7F]; /* 8 bits */
     y = 1 - a * r;
